foobar/preloader/requirements.txt
pandas
boto3
cassandra-driver